making useable functions from main
